Shark!
Created the sharks, which are dark lights moving around in intervals.
However, if the sharks are moving, the player (yellow) also blinks.
Must fix that. Additionally, I created a game over, which works for a
second (then the pink dot disappears). Lots of bugs, but we’re on the
right track!